[WIP] Initial rework of tedana.py
